Positions GET Problem
Took 51 seconds

+++ src/main/java/me/mrliam2614/facilitiscommons/utils/Positions.java
@@ -48,8 +48,8 @@ public class Positions {
         this.y = (Double) cs.get("coordinates.y", 0.0);
         this.z = (Double) cs.get("coordinates.z", 0.0);
         this.worldID = UUID.fromString((String) cs.get("world", Bukkit.getWorld(FacilitisCommons.getFacilitisCommons().getJavaPlugin().getServer().getWorlds().get(0).getUID().toString())));
-        this.pitch = (float) cs.get("view.pitch", 0f);
-        this.yaw = (float) cs.get("view.yaw", 0f);
+        this.pitch = ((Double) cs.get("view.pitch", 0f)).floatValue();
+        this.yaw = ((Double) cs.get("view.yaw", 0f)).floatValue();
     }
 
     public void save(ConfigurationSection cs){
